refactor: extract types from schema + use js file to define schema
This change moves the schema to a javascript file, this allows us to
update the schema without having to run a build to transpile typescript.
