+++ src/config/types.ts
@@ -1,0 +1,6 @@
+import type { z } from "zod";
+import type { ForkConfigSchema } from "./schema";
+
+export type ForkConfig = z.infer<typeof ForkConfigSchema>;
+
+export type Config = Partial<ForkConfig>;
