+++ src/process/commit.ts
@@ -1,7 +1,7 @@
 import { resolve } from "node:path";
 import { formatCommitMessage } from "../utils/format-commit-message";
 import { fileExists } from "../utils/file-state";
-import type { ForkConfig } from "../config/schema";
+import type { ForkConfig } from "../config/types";
 import type { FileState } from "../strategies/file-manager";
 import type { Logger } from "../utils/logger";
 import type { Git } from "../utils/git";
