+++ src/strategies/plain-text.ts
@@ -2,7 +2,7 @@ import { resolve } from "node:path";
 import { readFileSync, writeFileSync } from "node:fs";
 
 import { fileExists } from "../utils/file-state";
-import type { ForkConfig } from "../config/schema";
+import type { ForkConfig } from "../config/types";
 import type { Logger } from "../utils/logger";
 import type { FileState, IFileManager } from "./file-manager";
